https://github.com/RedRoverSchool/JenkinsQA_Java_2025_spring/issues/1837 RF_03.001 | Pipeline | PipelineProjectTest - refactoring PipelineProjectTest was deleted all tests were moved to PipelineTest
