+++ src/test/java/school/redrover/PipelineTest.java
@@ -6,11 +6,13 @@ import school.redrover.common.BaseTest;
 import school.redrover.page.HomePage;
 import school.redrover.page.error.ErrorPage;
 import school.redrover.page.newitem.NewItemPage;
+import school.redrover.page.pipeline.PipelineConfigurationPage;
 import school.redrover.page.pipeline.PipelineProjectPage;
 
 public class PipelineTest extends BaseTest {
 
     private static final String PROJECT_NAME = "PipelineProjectNameTest";
+    private static final String PIPELINE_DESCRIPTION = "Pipeline Project Name";
 
     @Test
     public void testCreateNewPipeline() {
@@ -40,17 +42,6 @@ public class PipelineTest extends BaseTest {
         Assert.assertEquals(error.getErrorText(), "No such job: " + copyFrom);
     }
 
-    @Test
-    public void testEmptyItemNamePOM() {
-        NewItemPage newItemPage = new HomePage(getDriver())
-                .clickCreateJob()
-                .selectPipeline();
-
-        Assert.assertEquals(newItemPage.getEmptyNameMessage(), "» This field cannot be empty, please enter a valid name");
-        Assert.assertFalse(newItemPage.isOkButtonEnabled());
-        Assert.assertEquals(newItemPage.getEmptyNameMessageColor(), "rgba(230, 0, 31, 1)");
-    }
-
     @Test(dependsOnMethods = "testCreateNewPipeline")
     public void testDisableProject() {
 
@@ -63,4 +54,77 @@ public class PipelineTest extends BaseTest {
         Assert.assertEquals(disabledProject.getProjectDisabledStatus(),"This project is currently disabled\n" +
                 "Enable");
     }
+
+    @Test
+    public void testCreateWithDescription() {
+        PipelineProjectPage pipelineProjectPage = new HomePage(getDriver())
+                .clickCreateJob()
+                .sendItemName(PROJECT_NAME)
+                .selectPipelineAndClickOk()
+                .sendDescription(PIPELINE_DESCRIPTION)
+                .clickSave();
+
+        Assert.assertEquals(pipelineProjectPage.getProjectName(), PROJECT_NAME);
+        Assert.assertEquals(pipelineProjectPage.getDescription(), PIPELINE_DESCRIPTION);
+    }
+
+    @Test(dependsOnMethods = "testCreateWithDescription")
+    public void testDeletePipeline() {
+        HomePage homePage = new HomePage(getDriver())
+                .clickOnJobInListOfItems(PROJECT_NAME, new PipelineProjectPage(getDriver()))
+                .deletePipeline();
+
+        Assert.assertEquals(homePage.getProjectNameList().size(), 0);
+    }
+
+    @Test
+    public void testDisableProjectErrorWhenCreating() {
+        PipelineConfigurationPage pipelineConfigurationPage = new HomePage(getDriver())
+                .clickNewItemOnLeftSidePanel()
+                .sendItemName(PROJECT_NAME)
+                .selectPipelineAndClickOk()
+                .switchToggle();
+
+        Assert.assertTrue(pipelineConfigurationPage.isToggleDisabled(), "The switch is not in an active state");
+    }
+
+    @Test
+    public void testEnableProject() {
+        PipelineConfigurationPage pipelineConfigurationPage = new HomePage(getDriver())
+                .clickNewItemOnLeftSidePanel()
+                .sendItemName(PROJECT_NAME)
+                .selectPipelineAndClickOk()
+                .switchToggle()
+                .clickSave()
+                .clickConfigure()
+                .switchToggle();
+
+        Assert.assertTrue(pipelineConfigurationPage.isToggleEnabled(), "The switch is turned on after performing the actions");
+    }
+
+    @Test
+    public void testCheckDefaultState() {
+        String statusToggleDefault = new HomePage(getDriver())
+                .clickNewItemOnLeftSidePanel()
+                .sendItemName(PROJECT_NAME)
+                .selectPipelineAndClickOk()
+                .checkStatusOnToggle();
+
+        Assert.assertEquals(statusToggleDefault, "Enabled");
+    }
+
+    @Test
+    public void testChangeStateNewPipeline() {
+        String statusToggleChange = new HomePage(getDriver())
+                .clickNewItemOnLeftSidePanel()
+                .sendItemName(PROJECT_NAME)
+                .selectPipelineAndClickOk()
+                .switchToggle()
+                .checkStatusOffToggle();
+
+        Assert.assertEquals(statusToggleChange, "Disabled");
+
+    }
+
+
 }
